aparte Authority controller aangemaakt en verplaatst vanuit UserController en endpoint in securityconfig aangepast
