+++ src/main/resources/data.sql
@@ -3,16 +3,16 @@ DELETE FROM users WHERE username IN ('Tamara', 'Ariel', 'Elsa', 'Jane', 'Belle',
 DELETE FROM authorities WHERE username IN ('Tamara', 'Ariel', 'Elsa', 'Jane', 'Belle', 'Aurora', 'Tiana', 'Sarabi', 'Nala', 'Moana');
 -- Insert users with encrypted passwords
 INSERT INTO users (username, password, email, apikey, enabled, name, gender, dob, autism_diagnoses, autism_diagnoses_year, profile_picture_url) VALUES
-('Tamara', '$2a$10$YJceXZTm4cGdln1W06362uq56.WOix.1DuPxXvHRdh5C3YZI7t9aG', 'Tamara.debeer@hotmail.com', 'random_api_key', true, 'Tamara', 'Female', '1991-07-06', 'Ja', 2020, 'http://localhost:1991/images/Tamara.jpg'),
-('Ariel', '$2a$10$ZdCa/52/Kf4pr0rZuBUf1ue8yKyYvWZikc8pi3HJzjlunlgNQNjEe', 'Ariel@ariel.nl', 'random_api_key', true, 'Ariel', 'Female', '2004-08-23', 'Ja', 2010, 'http://localhost:1991/images/Ariel.jpg'),
-('Elsa', '$2a$10$MJ3/vZj4u/LY0KlUeaBScu/bQByYHAMNDQ0869wCQRylZMn040OcG', 'Elsa@Elsa.nl', 'random_api_key', true, 'Elsa', 'Female', '2002-06-20', 'Ja', 2018, 'http://localhost:1991/images/Elsa.jpg'),
-('Jane', '$2a$10$1fymiRFhozuVvOXZgKYBw.VvaMaqKHpPbnfxMehw5FaYPqgXbZP.S', 'Jane@Jane.nl', 'random_api_key', true, 'Jane', 'Female', '2000-02-05', 'Ja', 2015, 'http://localhost:1991/images/Jane.jpg'),
-('Belle', '$2a$10$hx14PrRPWvPnHyslxlCMUOi3UcoawmqRZ79ZVGDTte12zf1JeQ6wi', 'Belle@belle.nl', 'random_api_key', true, 'Belle', 'Female', '1995-12-06', 'Ja', 2020, 'http://localhost:1991/images/Belle.jpg'),
-('Aurora', '$2a$10$Ro6ONJOXf2JNdvmS9lp4velG21PvtivhMPJavsrq/.KPVKQctNDCe', 'Aurora@aurora.nl', 'random_api_key', true, 'Aurora', 'Female', '1945-12-06', 'Ja', 1963, 'http://localhost:1991/images/Aurora.jpg'),
-('Tiana', '$2a$10$.4V27Ao/3KU7D1dfLUTbI.qV68IPoblsXb0HHSa8iAMk7enjJy7bi', 'Tiana@tiana.nl', 'random_api_key', true, 'Tiana', 'Female', '1955-09-15', 'Ja', 1980, 'http://localhost:1991/images/Tiana.jpg'),
-('Sarabi', '$2a$10$y2ufeazhi3FPGUBeh0aQC.IbR5QDRyUoRHZPsV.16NTVh5kfwQMeC', 'Sarabi@sarabi.nl', 'random_api_key', true, 'Sarabi', 'Female', '1988-02-09', 'Ja', 2019, 'http://localhost:1991/images/Sarabi.jpg'),
-('Nala', '$2a$10$1Jt3CFscaQ4s5jz.72g0tOetsmJsaCiC0FyL.igOgJq21YHbPDfi2', 'Nala@nala.nl', 'random_api_key', true, 'Nala', 'Female', '1993-11-11', 'Ja', 2022, 'http://localhost:1991/images/Nala.jpg'),
-('Moana', '$2a$10$vTcOZ0o8z6l1gGONioA3cOaOM8OlWQZfWuaN51jhhHsbUczJ3KuZ2', 'Moana@moana.nl', 'random_api_key', true, 'Moana', 'Female', '1980-05-28', 'Ja', 2000, 'http://localhost:1991/images/Moana.jpg');
+('Tamara', '$2a$10$6Fr2X6Okk0v5IPSQkn70S.wcipMfucdfnWQQ2Wheo6I1fZT2nAHNC', 'Tamara.debeer@hotmail.com', 'random_api_key', true, 'Tamara', 'Female', '1991-07-06', 'Ja', 2020, 'http://localhost:1991/images/Tamara.jpg'),
+('Ariel', '$2a$10$M4K.c5AEBX9gHDsXgFSc1.W3K6C8jAVuamz8U4WcD31XihlIaVCq.', 'Ariel@ariel.nl', 'random_api_key', true, 'Ariel', 'Female', '2004-08-23', 'Ja', 2010, 'http://localhost:1991/images/Ariel.jpg'),
+('Elsa', '$2a$10$SXM8MepqmAuCUARXq2FtWux2J0qCLHzwM9Q49H1JDJIKfIjc632MW', 'Elsa@Elsa.nl', 'random_api_key', true, 'Elsa', 'Female', '2002-06-20', 'Ja', 2018, 'http://localhost:1991/images/Elsa.jpg'),
+('Jane', '$2a$10$rWqHAtrVCH3E3couzjdeEuXUAbeT0KXJk/W9XXfxeHjLfPmM6w34u', 'Jane@Jane.nl', 'random_api_key', true, 'Jane', 'Female', '2000-02-05', 'Ja', 2015, 'http://localhost:1991/images/Jane.jpg'),
+('Belle', '$2a$10$XmYUDMQJE3emZqYF0KdCRe/5R6vR0Bot//ckuiTG/fPLzmU15hdOe', 'Belle@belle.nl', 'random_api_key', true, 'Belle', 'Female', '1995-12-06', 'Ja', 2020, 'http://localhost:1991/images/Belle.jpg'),
+('Aurora', '$2a$10$mOWvY4zuizdtUOsze8hG/ezkJ/BJBZaD0yvdNfHUSRE0josKlPbYe', 'Aurora@aurora.nl', 'random_api_key', true, 'Aurora', 'Female', '1945-12-06', 'Ja', 1963, 'http://localhost:1991/images/Aurora.jpg'),
+('Tiana', '$2a$10$pGJSrqNmxmhKBLwNzxrQz.A9x6vnukJE/BaihYlb55dXSNzCR2tEK', 'Tiana@tiana.nl', 'random_api_key', true, 'Tiana', 'Female', '1955-09-15', 'Ja', 1980, 'http://localhost:1991/images/Tiana.jpg'),
+('Sarabi', '$2a$10$ACf9deKzGpziz91NiCyLbui3UtTdL8zEWY0e4V0Yf6mOUC048T8xC', 'Sarabi@sarabi.nl', 'random_api_key', true, 'Sarabi', 'Female', '1988-02-09', 'Ja', 2019, 'http://localhost:1991/images/Sarabi.jpg'),
+('Nala', '$2a$10$sHbysdSezArbe3hjrIXgMOtiS9bLfpFte/OA.1Lkj1tBtvBwYdhDK', 'Nala@nala.nl', 'random_api_key', true, 'Nala', 'Female', '1993-11-11', 'Ja', 2022, 'http://localhost:1991/images/Nala.jpg'),
+('Moana', '$2a$10$S7CdrD0cNWNy0LWYVsrgEudRY4dl8rSn1dBI84sYZS6eP/ss6bDF6', 'Moana@moana.nl', 'random_api_key', true, 'Moana', 'Female', '1980-05-28', 'Ja', 2000, 'http://localhost:1991/images/Moana.jpg');
 -- Insert authorities for users
 INSERT INTO authorities (username, authority) VALUES
 ('Tamara', 'ROLE_ADMIN'),
